Refactor API routes into dedicated TOTp and local import modules
